eigrpd: Fix integer overflow possibility

+++ eigrpd/eigrp_topology.c
@@ -473,7 +473,7 @@ eigrp_topology_update_node_flags(struct eigrp_prefix_entry *dest)
 
   for (ALL_LIST_ELEMENTS_RO(dest->entries, node, entry))
     {
-      if ((entry->distance <= (uint64_t)(dest->distance*eigrp->variance)) &&
+      if (((uint64_t)entry->distance <= (uint64_t)(dest->distance*eigrp->variance)) &&
           entry->distance != EIGRP_MAX_METRIC) // is successor
         {
           entry->flags |= EIGRP_NEIGHBOR_ENTRY_SUCCESSOR_FLAG;
